Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    double a, area;
-    scanf("%lf", a);
+    float a, area;
+    scanf("%f", a);
     area = 3.14*a*a;
-    printf("Area: %.2lf", area);
+    printf("Area: %.2f", area);
     return 0; 
 }
